CIMIS Data Flags

List CIMIS data quality control flags.

cimis_flags(type = c("Severe", "Informative"), period = "Current")

Arguments

type
The type of data flag, i.e. "Severe" or "Informative".

period
The Time period that data was collected, i.e., "Current" or "Former" (pre-1995).

Value

a dataframe of data flags.

See also

Examples

cimis_flags()

#>    Flag       Class  Period
#> 1     I      Severe Current
#> 2     M      Severe Current
#> 3     S      Severe Current
#> 4     A Informative Current
#> 5     E Informative Current
#> 6     T Informative Current
#> 7     H Informative Current
#> 8     J Informative Current
#> 9     K Informative Current
#> 10    L Informative Current
#> 11    N Informative Current
#> 12    P Informative Current
#> 13    Q Informative Current
#> 14    R Informative Current
#> 15    Y Informative Current
#>                                                                                                                                              Description
#> 1                                                                                                                      Data value has no meaning/ignore.
#> 2                                                                                                                                               Missing.
#> 3                                                                                           Sensor is not in service or data is out of sensor threshold.
#> 4                                                                                                                                    Historical average.
#> 5                                                                                Historical average of one of the sensors used to calculate a parameter.
#> 6                                                                                Historical average of one of the sensors used to calculate a parameter.
#> 7  Daily data value flagged when corresponding hourly data is flagged M, Q, and S. This flag is not set when the corresponding hourly data is flagged N.
#> 8                                                                                           Monthly data value flagged when some daily data are missing.
#> 9                                                                                         Monthly data value flagged when daily data values are flagged.
#> 10                                                                     Monthly data value flagged when some daily data are missing and some are flagged.
#> 11                                                                                                          Data value is not collected by this station.
#> 12                                                                                                                                 Quality test pending.
#> 13       All quality control could not be performed because a comparison sensor is severe. This flag is not set when the comparison sensor is flagged N.
#> 14                                                                                                                 Data is far out of historical limits.
#> 15                                                                                                          Data is moderately out of historical limits.

cimis_flags("Informative")

#>    Flag       Class  Period
#> 1     A Informative Current
#> 2     E Informative Current
#> 3     T Informative Current
#> 4     H Informative Current
#> 5     J Informative Current
#> 6     K Informative Current
#> 7     L Informative Current
#> 8     N Informative Current
#> 9     P Informative Current
#> 10    Q Informative Current
#> 11    R Informative Current
#> 12    Y Informative Current
#>                                                                                                                                              Description
#> 1                                                                                                                                    Historical average.
#> 2                                                                                Historical average of one of the sensors used to calculate a parameter.
#> 3                                                                                Historical average of one of the sensors used to calculate a parameter.
#> 4  Daily data value flagged when corresponding hourly data is flagged M, Q, and S. This flag is not set when the corresponding hourly data is flagged N.
#> 5                                                                                           Monthly data value flagged when some daily data are missing.
#> 6                                                                                         Monthly data value flagged when daily data values are flagged.
#> 7                                                                      Monthly data value flagged when some daily data are missing and some are flagged.
#> 8                                                                                                           Data value is not collected by this station.
#> 9                                                                                                                                  Quality test pending.
#> 10       All quality control could not be performed because a comparison sensor is severe. This flag is not set when the comparison sensor is flagged N.
#> 11                                                                                                                 Data is far out of historical limits.
#> 12                                                                                                          Data is moderately out of historical limits.

cimis_flags("Severe", period = "Former")

#>   Flag  Class Period
#> 1    I Severe Former
#> 2  N/A Severe Former
#> 3    S Severe Former
#>                                                    Description
#> 1                                  Data has no meaning/ignore.
#> 2                                       Data is not available.
#> 3 Sensor is not in service or data is out of sensor threshold.